requirements.txt
# agent_web 依賴（含 LangChain RAG）
fastapi
uvicorn
python-multipart
python-dotenv
openai
requests
pydantic

# LangChain RAG
langchain
langchain-openai
langchain-community
langchain-chroma
